+++ custom_components/ember_mug/sensor.py
@@ -108,7 +108,7 @@ class EmberMugStateSensor(EmberMugSensor):
             "firmware_info": data.firmware,
             "raw_state": data.liquid_state,
         }
-        if data.model.include_extra:
+        if data.debug:
             attrs |= {
                 "date_time_zone": data.date_time_zone,
                 "udsk": data.udsk,
@@ -180,7 +180,7 @@ class EmberMugBatterySensor(EmberMugSensor):
         attrs = {
             ATTR_BATTERY_CHARGING: data.battery.on_charging_base if data.battery else None,
         }
-        if ATTR_BATTERY_VOLTAGE in data.model.attribute_labels:
+        if ATTR_BATTERY_VOLTAGE in data.model_info.update_attributes:
             attrs[ATTR_BATTERY_VOLTAGE] = data.battery_voltage
         return attrs | super().extra_state_attributes